Added basic implementation spectrogram plot
Now the fft is not implemented, a log10 version of the time signal is
adapted and displayed in the spectrogram. It seems works, from now must
be implemented the fft.
Spectrogram is instantiated from mainwindow. Is this right?
Now a QwtRasterData raster data is used, a better solution could be use
QwtRasterMatrixData